Changed variables to lower variable types

+++ Commission/D_Functions.ino
@@ -4,7 +4,7 @@ void StopSystem() {
   digitalWrite(OzoneEmitter, HIGH);
 
   ozoneReading = Ozone.readOzoneData(COLLECT_NUMBER) / 1000.0;
-  
+
   //turn on filter while ozone is high
   if (ozoneReading > safeValue) {
     displayCleaningMessage();
@@ -15,7 +15,7 @@ void StopSystem() {
     //if paused is used, and start is ran, it will restart the value of the timer
     timer.pause();
     digitalWrite(CarbonFilter, HIGH);
-    
+
     //Stop the timer task which is running this stop system function
     displayStartMessage();
     //Serial.println("Ozone Regulated");
@@ -56,16 +56,16 @@ void displayCleaningMessage() {
 }
 
 void displayTimeElapsed() {
-  float timeElapsed = (timer.read() / 1000.0) / 60.0;
+  uint32_t timeElapsed = (timer.read() / 1000.0) / 60.0;
   int timeElapsedWhole = static_cast<int>(timeElapsed);
 
   char secondsCount[2];
-  float seconds = (timeElapsed - timeElapsedWhole) * 60;
+  uint8_t seconds = (timeElapsed - timeElapsedWhole) * 60;
   dtostrf(seconds, 2, 0, secondsCount);
   if (seconds <= 9.9999) {
     secondsCount[0] = '0';  
   }
-
+  
   //display time elapsed
   lcd.setCursor(3, 1);
   lcd.print("Time: ");
